+++ modules/canvaskit/CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    but can be an array, a TypedArray, or a MallocObj.
  - `CanvasKit.*Builder` have all been removed. Clients should use Malloc instead.
 
+### Removed
+ - `CanvasKit.Shader.MakeLerp`, the same effect can be easily generated with `RuntimeEffect`
+
 ### Known Bugs
  - On legacy (non-ANGLE) SwiftShader, certain paths that require tessellation may not be drawn
    correctly when using a WebGL-backed surface. (skbug.com/11965)